+++ packages/api-client/src/views/Request/RequestSection/RequestAuth/OAuthScopesInput.vue
@@ -79,20 +79,26 @@ const deselectAllScopes = () => {
             {{ Object.keys(flow?.scopes ?? {}).length || 0 }}
           </div>
           <div class="flex items-center gap-1.75">
+            <!-- Deselect All -->
             <ScalarButton
               v-if="allScopesSelected"
-              class="text-c-3 hover:bg-b-2 hover:text-c-1 rounded px-1.5"
+              class="pr-0.75 pl-1 transition-none"
               size="sm"
+              variant="ghost"
               @click.stop="deselectAllScopes">
               Deselect All
             </ScalarButton>
+
+            <!-- Select All -->
             <ScalarButton
               v-if="!allScopesSelected"
-              class="text-c-3 hover:bg-b-2 hover:text-c-1 rounded px-1.5"
+              class="pr-0.75 pl-1 transition-none"
               size="sm"
+              variant="ghost"
               @click.stop="selectAllScopes">
               Select All
             </ScalarButton>
+
             <ScalarIcon
               class="text-c-3 group-hover/scopes-accordion:text-c-2"
               :icon="open ? 'ChevronDown' : 'ChevronRight'"