+++ integrations/aspire/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @scalar/aspire
 
+## 0.4.2
+
+### Patch Changes
+
+- Updated dependencies [50032be]
+- Updated dependencies [88385b1]
+  - @scalar/api-reference@1.34.6
+
+## 0.4.1
+
+### Patch Changes
+
+- Updated dependencies [86d29c1]
+  - @scalar/api-reference@1.34.5
+
+## 0.4.0
+
+### Minor Changes
+
+- e4279e4: feat: support HTTPS
+- 508c04a: feat: use environment callback instead of hooks
+
+### Patch Changes
+
+- 2f05f7d: feat: support multiple scalar resources
+- a995f15: feat: support async configuration
+- Updated dependencies [bbef120]
+- Updated dependencies [24883ab]
+- Updated dependencies [a1429ca]
+- Updated dependencies [e499aee]
+- Updated dependencies [845e850]
+- Updated dependencies [66b18fc]
+- Updated dependencies [6eb7c4f]
+- Updated dependencies [4e06829]
+- Updated dependencies [5f022b5]
+- Updated dependencies [2b89e6f]
+- Updated dependencies [6a88108]
+- Updated dependencies [c838a3f]
+- Updated dependencies [75d84ed]
+- Updated dependencies [929df42]
+- Updated dependencies [0804cfc]
+  - @scalar/api-reference@1.34.4
+
 ## 0.3.3
 
 ### Patch Changes