+++ packages/api-client/CHANGELOG.md
@@ -1,5 +1,69 @@
 # @scalar/api-client
 
+## 2.5.26
+
+### Patch Changes
+
+- Updated dependencies [50032be]
+  - @scalar/types@0.2.13
+  - @scalar/oas-utils@0.4.22
+  - @scalar/openapi-parser@0.20.1
+  - @scalar/snippetz@0.4.7
+  - @scalar/themes@0.13.14
+  - @scalar/components@0.14.27
+  - @scalar/postman-to-openapi@0.3.25
+  - @scalar/import@0.4.19
+  - @scalar/use-codemirror@0.12.28
+
+## 2.5.25
+
+### Patch Changes
+
+- Updated dependencies [86d29c1]
+- Updated dependencies [e2272a1]
+- Updated dependencies [b6f4b60]
+  - @scalar/snippetz@0.4.6
+  - @scalar/oas-utils@0.4.21
+  - @scalar/components@0.14.26
+  - @scalar/postman-to-openapi@0.3.24
+  - @scalar/use-codemirror@0.12.27
+
+## 2.5.24
+
+### Patch Changes
+
+- 172e397: feat(components): add hotkey component and tooltip support
+- 5f022b5: feat: use registry.scalar.com URLs as example
+- 0804cfc: eat(components): Add ScalarMarkdown summary component
+- Updated dependencies [bbef120]
+- Updated dependencies [a1429ca]
+- Updated dependencies [d4adeba]
+- Updated dependencies [172e397]
+- Updated dependencies [66b18fc]
+- Updated dependencies [5f022b5]
+- Updated dependencies [d8b8a5b]
+- Updated dependencies [6c24be9]
+- Updated dependencies [6a88108]
+- Updated dependencies [c418e92]
+- Updated dependencies [638120c]
+- Updated dependencies [0804cfc]
+  - @scalar/oas-utils@0.4.20
+  - @scalar/openapi-parser@0.20.0
+  - @scalar/components@0.14.25
+  - @scalar/helpers@0.0.8
+  - @scalar/themes@0.13.13
+  - @scalar/types@0.2.12
+  - @scalar/postman-to-openapi@0.3.23
+  - @scalar/import@0.4.18
+  - @scalar/use-codemirror@0.12.26
+  - @scalar/draggable@0.2.0
+  - @scalar/icons@0.4.7
+  - @scalar/object-utils@1.2.4
+  - @scalar/openapi-types@0.3.7
+  - @scalar/snippetz@0.4.5
+  - @scalar/use-hooks@0.2.4
+  - @scalar/use-toasts@0.8.0
+
 ## 2.5.23
 
 ### Patch Changes