+++ packages/import/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @scalar/import
 
+## 0.4.19
+
+### Patch Changes
+
+- @scalar/openapi-parser@0.20.1
+
+## 0.4.18
+
+### Patch Changes
+
+- Updated dependencies [d4adeba]
+- Updated dependencies [66b18fc]
+- Updated dependencies [5f022b5]
+- Updated dependencies [c418e92]
+  - @scalar/openapi-parser@0.20.0
+  - @scalar/helpers@0.0.8
+
 ## 0.4.17
 
 ### Patch Changes