+++ packages/pre-post-request-scripts/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @scalar/scripts
 
+## 0.0.28
+
+### Patch Changes
+
+- @scalar/oas-utils@0.4.22
+- @scalar/components@0.14.27
+
+## 0.0.27
+
+### Patch Changes
+
+- Updated dependencies [e2272a1]
+- Updated dependencies [b6f4b60]
+  - @scalar/oas-utils@0.4.21
+  - @scalar/components@0.14.26
+
+## 0.0.26
+
+### Patch Changes
+
+- Updated dependencies [bbef120]
+- Updated dependencies [a1429ca]
+- Updated dependencies [172e397]
+- Updated dependencies [66b18fc]
+- Updated dependencies [5f022b5]
+- Updated dependencies [d8b8a5b]
+- Updated dependencies [6c24be9]
+- Updated dependencies [638120c]
+- Updated dependencies [0804cfc]
+  - @scalar/oas-utils@0.4.20
+  - @scalar/components@0.14.25
+
 ## 0.0.25
 
 ### Patch Changes