+++ packages/snippetz/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @scalar/snippetz
 
+## 0.4.7
+
+### Patch Changes
+
+- Updated dependencies [50032be]
+  - @scalar/types@0.2.13
+
+## 0.4.6
+
+### Patch Changes
+
+- 86d29c1: feat(snippetz): improve indentation handling in objectToString
+
+## 0.4.5
+
+### Patch Changes
+
+- Updated dependencies [6a88108]
+  - @scalar/types@0.2.12
+
 ## 0.4.4
 
 ### Patch Changes