+++ projects/scalar-app/CHANGELOG.md
@@ -1,5 +1,36 @@
 # scalar-app
 
+## 0.1.224
+
+### Patch Changes
+
+- @scalar/api-client@2.5.26
+- @scalar/themes@0.13.14
+- @scalar/components@0.14.27
+- @scalar/import@0.4.19
+
+## 0.1.223
+
+### Patch Changes
+
+- @scalar/api-client@2.5.25
+- @scalar/components@0.14.26
+
+## 0.1.222
+
+### Patch Changes
+
+- Updated dependencies [172e397]
+- Updated dependencies [5f022b5]
+- Updated dependencies [d8b8a5b]
+- Updated dependencies [6c24be9]
+- Updated dependencies [638120c]
+- Updated dependencies [0804cfc]
+  - @scalar/api-client@2.5.24
+  - @scalar/components@0.14.25
+  - @scalar/themes@0.13.13
+  - @scalar/import@0.4.18
+
 ## 0.1.221
 
 ### Patch Changes